phishingsystem/monitoring_pipeline/requirements.txt
azure-functions
azure-storage-blob
pymongo
pandas
numpy
requests
python-dotenv